Enable race detector for top-level unit tests (#11207)
