+++ bootstrap/kubeadm/types/upstreamv1beta4/conversion_test.go
@@ -1,3 +1,5 @@
+//go:build !race
+
 /*
 Copyright 2021 The Kubernetes Authors.
 
@@ -18,18 +20,18 @@ package upstreamv1beta4
 
 import (
 	"testing"
-	"time"
 
 	fuzz "github.com/google/gofuzz"
 	. "github.com/onsi/gomega"
 	"k8s.io/apimachinery/pkg/api/apitesting/fuzzer"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	runtimeserializer "k8s.io/apimachinery/pkg/runtime/serializer"
 
 	bootstrapv1 "sigs.k8s.io/cluster-api/bootstrap/kubeadm/api/v1beta1"
 	utilconversion "sigs.k8s.io/cluster-api/util/conversion"
 )
+
+// Test is disabled when the race detector is enabled (via "//go:build !race" above) because otherwise the fuzz tests would just time out.
 
 func TestFuzzyConversion(t *testing.T) {
 	g := NewWithT(t)
@@ -143,42 +145,4 @@ func bootstrapv1JoinConfigurationFuzzer(obj *bootstrapv1.JoinConfiguration, c fu
 	if obj.Discovery.File != nil {
 		obj.Discovery.File.KubeConfig = nil
 	}
-}
-
-func TestTimeoutForControlPlaneMigration(t *testing.T) {
-	timeout := metav1.Duration{Duration: 10 * time.Second}
-	t.Run("from ClusterConfiguration to InitConfiguration and back", func(t *testing.T) {
-		g := NewWithT(t)
-
-		clusterConfiguration := &bootstrapv1.ClusterConfiguration{
-			APIServer: bootstrapv1.APIServer{TimeoutForControlPlane: &timeout},
-		}
-
-		initConfiguration := &InitConfiguration{}
-		err := initConfiguration.ConvertFromClusterConfiguration(clusterConfiguration)
-		g.Expect(err).ToNot(HaveOccurred())
-		g.Expect(initConfiguration.Timeouts.ControlPlaneComponentHealthCheck).To(Equal(&timeout))
-
-		clusterConfiguration = &bootstrapv1.ClusterConfiguration{}
-		err = initConfiguration.ConvertToClusterConfiguration(clusterConfiguration)
-		g.Expect(err).ToNot(HaveOccurred())
-		g.Expect(clusterConfiguration.APIServer.TimeoutForControlPlane).To(Equal(&timeout))
-	})
-	t.Run("from ClusterConfiguration to JoinConfiguration and back", func(t *testing.T) {
-		g := NewWithT(t)
-
-		clusterConfiguration := &bootstrapv1.ClusterConfiguration{
-			APIServer: bootstrapv1.APIServer{TimeoutForControlPlane: &timeout},
-		}
-
-		joinConfiguration := &JoinConfiguration{}
-		err := joinConfiguration.ConvertFromClusterConfiguration(clusterConfiguration)
-		g.Expect(err).ToNot(HaveOccurred())
-		g.Expect(joinConfiguration.Timeouts.ControlPlaneComponentHealthCheck).To(Equal(&timeout))
-
-		clusterConfiguration = &bootstrapv1.ClusterConfiguration{}
-		err = joinConfiguration.ConvertToClusterConfiguration(clusterConfiguration)
-		g.Expect(err).ToNot(HaveOccurred())
-		g.Expect(clusterConfiguration.APIServer.TimeoutForControlPlane).To(Equal(&timeout))
-	})
 }
