+++ internal/apis/bootstrap/kubeadm/v1alpha4/conversion_test.go
@@ -1,3 +1,5 @@
+//go:build !race
+
 /*
 Copyright 2021 The Kubernetes Authors.
 
@@ -31,6 +33,8 @@ const (
 	fakeID     = "abcdef"
 	fakeSecret = "abcdef0123456789"
 )
+
+// Test is disabled when the race detector is enabled (via "//go:build !race" above) because otherwise the fuzz tests would just time out.
 
 func TestFuzzyConversion(t *testing.T) {
 	t.Run("for KubeadmConfig", utilconversion.FuzzTestFunc(utilconversion.FuzzTestFuncInput{
