+++ internal/apis/core/v1alpha3/conversion_test.go
@@ -1,3 +1,5 @@
+//go:build !race
+
 /*
 Copyright 2020 The Kubernetes Authors.
 
@@ -28,6 +30,8 @@ import (
 	clusterv1 "sigs.k8s.io/cluster-api/api/v1beta1"
 	utilconversion "sigs.k8s.io/cluster-api/util/conversion"
 )
+
+// Test is disabled when the race detector is enabled (via "//go:build !race" above) because otherwise the fuzz tests would just time out.
 
 func TestFuzzyConversion(t *testing.T) {
 	t.Run("for Cluster", utilconversion.FuzzTestFunc(utilconversion.FuzzTestFuncInput{
